// 02-state-and-reactivity/exercise/src/types.ts
export type IngredientType = 'bun' | 'sauce' | 'main';

export interface Ingredient {
  id: string;
  name: string;
  type: IngredientType;
  price: number;
  image: string;
}

export type CatalogStatus = 'loading' | 'ready';

export interface BurgerItem {
  uid: string;
  ingredient: Ingredient;
}